requirements.txt
streamlit>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0 
sqlalchemy>=2.0.0 